# docs/swagger.yaml
definitions:
  structtypes.Addapireq:
    properties:
      bodyrecodes:
        description: body结构
        type: string
      bodytype:
        description: body的类型
        type: string
      headcode:
        description: 请求头部结构
        type: string
      isDeprecated:
        description: 弃用状态
        enum:
        - 1
        - 2
        type: integer
      path:
        description: 路径
        type: string
      pathcode:
        description: path结构
        type: string
      projectname:
        description: 项目名
        type: string
      returncodes:
        description: 返回结构
        type: string
      returntype:
        description: 返回类型
        type: string
      tokens:
        description: 用户token
        type: string
      types:
        description: 请求类型
        enum:
        - get
        - post
        type: string
      username:
        description: 用户名
        type: string
      words:
        description: 注释
        type: string
    required:
    - bodyrecodes
    - bodytype
    - headcode
    - isDeprecated
    - path
    - pathcode
    - projectname
    - returncodes
    - returntype
    - tokens
    - types
    - username
    - words
    type: object
  structtypes.Createproject:
    properties:
      name:
        description: 名字
        type: string
      tokens:
        description: 用户token
        type: string
      username:
        description: 用户名
        type: string
    required:
    - name
    - tokens
    - username
    type: object
  structtypes.Delapi:
    properties:
      apiid:
        type: integer
      projectname:
        description: 项目名
        type: string
      tokens:
        description: 用户token
        type: string
      username:
        description: 用户名
        type: string
    required:
    - apiid
    - projectname
    - tokens
    - username
    type: object
  structtypes.Delproject:
    properties:
      name:
        description: 名字
        type: string
      tokens:
        description: 用户token
        type: string
      username:
        description: 用户名
        type: string
    required:
    - name
    - tokens
    - username
    type: object
  structtypes.GetProjectlist:
    properties:
      tokens:
        description: 用户token
        type: string
      username:
        description: 用户名
        type: string
    required:
    - tokens
    - username
    type: object
  structtypes.Getapi:
    properties:
      apiid:
        type: integer
      projectname:
        description: 项目名
        type: string
      tokens:
        description: 用户token
        type: string
      username:
        description: 用户名
        type: string
    required:
    - apiid
    - projectname
    - tokens
    - username
    type: object
  structtypes.Getapilist:
    properties:
      projectname:
        description: 项目名
        type: string
      tokens:
        description: 用户token
        type: string
      username:
        description: 用户名
        type: string
    required:
    - projectname
    - tokens
    - username
    type: object
  structtypes.Getapilistres:
    properties:
      id:
        description: id
        type: integer
      isDeprecated:
        description: 弃用状态 1表示正常 2表示弃用
        type: integer
      path:
        description: 路径
        type: string
      types:
        description: 请求类型
        type: string
    type: object
  structtypes.Getapires:
    properties:
      bodyrecodes:
        description: body结构
        type: string
      bodytype:
        description: body的类型
        type: string
      headcode:
        description: 请求头部结构
        type: string
      isDeprecated:
        description: 弃用状态 1表示正常 2表示弃用
        type: integer
      path:
        description: 路径
        type: string
      pathcode:
        description: path结构
        type: string
      returncodes:
        description: 返回结构
        type: string
      returntype:
        description: 返回类型
        type: string
      types:
        description: 请求类型
        type: string
      words:
        description: 注释
        type: string
    type: object
  structtypes.Getpermissionlist:
    properties:
      projectname:
        description: 项目名
        type: string
      tokens:
        description: 用户token
        type: string
      username:
        description: 用户名
        type: string
    required:
    - projectname
    - tokens
    - username
    type: object
  structtypes.Getpermissionlistres:
    properties:
      levels:
        type: integer
      userID:
        type: integer
      username:
        type: string
    type: object
  structtypes.Getprojectlistres:
    properties:
      apinum:
        type: integer
      creater:
        type: string
      projectid:
        type: integer
      projectname:
        type: string
    type: object
  structtypes.JSONResult:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
  structtypes.Setpermission:
    properties:
      adminnames:
        description: 管理员
        type: string
      editnames:
        description: 编辑者
        type: string
      guestnames:
        description: 访客
        type: string
      projectname:
        description: 项目名
        type: string
      tokens:
        description: 用户token
        type: string
      username:
        description: 用户名
        type: string
    required:
    - projectname
    - tokens
    - username
    type: object
  structtypes.UserLogin:
    properties:
      language:
        type: string
      password:
        type: string
      username:
        type: string
    required:
    - language
    - password
    - username
    type: object
  structtypes.UserRegReq:
    properties:
      email:
        type: string
      language:
        type: string
      password:
        type: string
      username:
        type: string
    required:
    - email
    - language
    - password
    - username
    type: object
  structtypes.Usersearchres:
    properties:
      id:
        type: integer
      username:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  title: getneko api doc
  version: "1.0"
paths:
  /v1/addapi:
    post:
      consumes:
      - application/json
      description: 参数校验慢慢完善
      parameters:
      - description: api info
        in: body
        name: UserLogin
        required: true
        schema:
          $ref: '#/definitions/structtypes.Addapireq'
      produces:
      - application/json
      responses:
        "200":
          description: desc
          schema:
            allOf:
            - $ref: '#/definitions/structtypes.JSONResult'
            - properties:
                data:
                  type: string
              type: object
      summary: 创建api
      tags:
      - api记录操作
  /v1/chacklowversion:
    get:
      consumes:
      - application/json
      description: 比较客户端与服务器的版本号
      produces:
      - application/json
      responses:
        "200":
          description: desc
          schema:
            allOf:
            - $ref: '#/definitions/structtypes.JSONResult'
            - properties:
                data:
                  type: string
              type: object
      summary: 获取客户端最低版本
      tags:
      - 客户端信息
  /v1/createproject:
    post:
      consumes:
      - application/json
      description: 目前暂时没有做权限
      parameters:
      - description: project info
        in: body
        name: UserLogin
        required: true
        schema:
          $ref: '#/definitions/structtypes.Createproject'
      produces:
      - application/json
      responses:
        "200":
          description: desc
          schema:
            allOf:
            - $ref: '#/definitions/structtypes.JSONResult'
            - properties:
                data:
                  type: string
              type: object
      summary: 创建项目
      tags:
      - 项目操作
  /v1/delapi:
    post:
      consumes:
      - application/json
      description: 没有说明
      parameters:
      - description: api info
        in: body
        name: UserLogin
        required: true
        schema:
          $ref: '#/definitions/structtypes.Delapi'
      produces:
      - application/json
      responses:
        "200":
          description: desc
          schema:
            allOf:
            - $ref: '#/definitions/structtypes.JSONResult'
            - properties:
                data:
                  type: string
              type: object
      summary: 删除api
      tags:
      - api记录操作
  /v1/delproject:
    post:
      consumes:
      - application/json
      description: 只要创建者才能删除
      parameters:
      - description: project info
        in: body
        name: UserLogin
        required: true
        schema:
          $ref: '#/definitions/structtypes.Delproject'
      produces:
      - application/json
      responses:
        "200":
          description: desc
          schema:
            allOf:
            - $ref: '#/definitions/structtypes.JSONResult'
            - properties:
                data:
                  type: string
              type: object
      summary: 删除项目
      tags:
      - 项目操作
  /v1/getapibyid:
    post:
      consumes:
      - application/json
      description: 目前没有需要注意的
      parameters:
      - description: api info
        in: body
        name: UserLogin
        required: true
        schema:
          $ref: '#/definitions/structtypes.Getapi'
      produces:
      - application/json
      responses:
        "200":
          description: desc
          schema:
            allOf:
            - $ref: '#/definitions/structtypes.JSONResult'
            - properties:
                data:
                  $ref: '#/definitions/structtypes.Getapires'
              type: object
      summary: 通过id获取api
      tags:
      - api记录操作
  /v1/getapilist:
    post:
      consumes:
      - application/json
      description: 目前没有需要注意的
      parameters:
      - description: api info
        in: body
        name: UserLogin
        required: true
        schema:
          $ref: '#/definitions/structtypes.Getapilist'
      produces:
      - application/json
      responses:
        "200":
          description: desc
          schema:
            allOf:
            - $ref: '#/definitions/structtypes.JSONResult'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/structtypes.Getapilistres'
                  type: array
              type: object
      summary: 获取api列表
      tags:
      - api记录操作
  /v1/getpermissionlist:
    post:
      consumes:
      - application/json
      description: 需要管理员权限
      parameters:
      - description: project info
        in: body
        name: UserLogin
        required: true
        schema:
          $ref: '#/definitions/structtypes.Getpermissionlist'
      produces:
      - application/json
      responses:
        "200":
          description: desc
          schema:
            allOf:
            - $ref: '#/definitions/structtypes.JSONResult'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/structtypes.Getpermissionlistres'
                  type: array
              type: object
      summary: 获取项目权限表
      tags:
      - 权限操作
  /v1/getprojectlist:
    post:
      consumes:
      - application/json
      description: 项目选择页面用的
      parameters:
      - description: api info
        in: body
        name: UserLogin
        required: true
        schema:
          $ref: '#/definitions/structtypes.GetProjectlist'
      produces:
      - application/json
      responses:
        "200":
          description: desc
          schema:
            allOf:
            - $ref: '#/definitions/structtypes.JSONResult'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/structtypes.Getprojectlistres'
                  type: array
              type: object
      summary: 查询项目列表
      tags:
      - 项目操作
  /v1/setpermission:
    post:
      consumes:
      - application/json
      description: 需要管理员权限
      parameters:
      - description: project info
        in: body
        name: UserLogin
        required: true
        schema:
          $ref: '#/definitions/structtypes.Setpermission'
      produces:
      - application/json
      responses:
        "200":
          description: desc
          schema:
            allOf:
            - $ref: '#/definitions/structtypes.JSONResult'
            - properties:
                data:
                  type: string
              type: object
      summary: 设置权限
      tags:
      - 权限操作
  /v1/userlogin:
    post:
      consumes:
      - application/json
      description: 用户登陆api
      parameters:
      - description: user info
        in: body
        name: UserLogin
        required: true
        schema:
          $ref: '#/definitions/structtypes.UserLogin'
      produces:
      - application/json
      responses:
        "200":
          description: desc
          schema:
            allOf:
            - $ref: '#/definitions/structtypes.JSONResult'
            - properties:
                data:
                  type: string
              type: object
      summary: 用户登陆
      tags:
      - 用户操作
  /v1/userreg:
    post:
      consumes:
      - application/json
      description: 用户注册api
      parameters:
      - description: user info
        in: body
        name: UserRegReq
        required: true
        schema:
          $ref: '#/definitions/structtypes.UserRegReq'
      produces:
      - application/json
      responses:
        "200":
          description: desc
          schema:
            allOf:
            - $ref: '#/definitions/structtypes.JSONResult'
            - properties:
                data:
                  type: string
              type: object
      summary: 用户注册
      tags:
      - 用户操作
  /v1/usersearch:
    get:
      consumes:
      - application/json
      description: 无需鉴权
      parameters:
      - in: query
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: desc
          schema:
            allOf:
            - $ref: '#/definitions/structtypes.JSONResult'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/structtypes.Usersearchres'
                  type: array
              type: object
      summary: 模糊查询用户
      tags:
      - 用户操作
swagger: "2.0"